Configuration file functional
The configuration file in /etc/ftposcd.conf is now used to configure the remote hosts IP and OSC port for its OSC server.  Also, the local OSC port for incoming OSC messages to be sent into the fishman triple play, and the Command and Control port, for sending in OSC which controls ftposcd itelf.  The cac port and function may very likely become depricated, as ftposcd is intended to be configured then ran as a daemon/service on a system.