Removed Command and Control concept, and work on incoming OSC to MIDI conversion
There is no need for the additional Command and Control OSC sever, as ftposcd is intended to be a service controlled by the init system (Systemd, OpenRC, etc...).  This commit contain further work on the incoming OSC server, which is used for converting incoming OSC into MIDI messages that control the functionality of the Fishman Triple Play.